restructure services in cache module
