+++ client/src/app/+device-management/humidity-setup-overview/humidity-setup-overview.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from "@angular/core";
 import {IHumidityDevice} from "../../../../../server/entities/device.interface";
-import {HumidityDeviceCacheService} from "../../cache/humidity-device.cache.service";
+import {HumidityDeviceCacheService} from "../../cache/service/humidity-device.cache.service";
 import {ReplaySubject, Observable} from "rxjs";
 import {List} from "immutable";
