+++ client/src/app/+device-management/humidity-setup-overview/humiditydevice-details/humiditydevice-details.component.ts
@@ -3,7 +3,7 @@ import {Router, ActivatedRoute} from "@angular/router";
 import {IHumidityDevice} from "../../../../../../server/entities/device.interface";
 import {Subscription} from "rxjs";
 import {Port, portName} from "../../../../../../server/hardware/port-map";
-import {HumidityDeviceCacheService} from "../../../cache/humidity-device.cache.service";
+import {HumidityDeviceCacheService} from "../../../cache/service/humidity-device.cache.service";
 
 @Component({
   selector: 'app-humiditydevice-details',
