+++ client/src/app/+device-management/service/analog-port.service.ts
@@ -3,8 +3,8 @@ import {ReplaySubject, Observable, Subscription} from "rxjs";
 import {List} from "immutable";
 import {ITemperatureDevice} from "../../../../../server/entities/device.interface";
 import {Port, analogInputs} from "../../../../../server/hardware/port-map";
-import {TemperatureDeviceCacheService} from "../../cache/temperature-device.cache.service";
-import {HumidityDeviceCacheService} from "../../cache/humidity-device.cache.service";
+import {TemperatureDeviceCacheService} from "../../cache/service/temperature-device.cache.service";
+import {HumidityDeviceCacheService} from "../../cache/service/humidity-device.cache.service";
 
 @Injectable()
 export class AnalogPortService {
