+++ client/src/app/+device-management/temperature-setup-overview/temperature-setup-overview.component.ts
@@ -2,7 +2,7 @@ import {Component, OnInit} from "@angular/core";
 import {ITemperatureDevice} from "../../../../../server/entities/device.interface";
 import {ReplaySubject, Observable} from "rxjs";
 import {List} from "immutable";
-import {TemperatureDeviceCacheService} from "../../cache/temperature-device.cache.service";
+import {TemperatureDeviceCacheService} from "../../cache/service/temperature-device.cache.service";
 
 @Component({
   selector: 'app-temperature-setup-overview',
