+++ client/src/app/cache/cache.module.ts
@@ -1,9 +1,9 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {TemperatureDeviceCacheService} from "./temperature-device.cache.service";
-import {BlindsDeviceCacheService} from "./blinds-device.cache.service";
-import {HumidityDeviceCacheService} from "./humidity-device.cache.service";
-import {UserCacheService} from "./user.cache.service";
+import {TemperatureDeviceCacheService} from "./service/temperature-device.cache.service";
+import {BlindsDeviceCacheService} from "./service/blinds-device.cache.service";
+import {HumidityDeviceCacheService} from "./service/humidity-device.cache.service";
+import {UserCacheService} from "./service/user.cache.service";
 
 @NgModule({
   imports: [
