+++ client/src/app/cache/service/blinds-device.cache.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from "@angular/core";
 import {AuthHttp} from "angular2-jwt";
 import {GenericeCacheService} from "./generic.cache.service";
-import {ClientSocketService} from "../remote/client-socket.service";
-import {NotificationService} from "../notification/notification.service";
-import {IBlindsDevice} from "../../../../server/entities/device.interface";
+import {ClientSocketService} from "../../remote/client-socket.service";
+import {NotificationService} from "../../notification/notification.service";
+import {IBlindsDevice} from "../../../../../server/entities/device.interface";
 
 @Injectable()
 export class BlindsDeviceCacheService extends GenericeCacheService<IBlindsDevice> {
