+++ client/src/app/cache/service/humidity-device.cache.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from "@angular/core";
 import {AuthHttp} from "angular2-jwt";
 import {GenericeCacheService} from "./generic.cache.service";
-import {ClientSocketService} from "../remote/client-socket.service";
-import {NotificationService} from "../notification/notification.service";
-import {IHumidityDevice} from "../../../../server/entities/device.interface";
+import {ClientSocketService} from "../../remote/client-socket.service";
+import {NotificationService} from "../../notification/notification.service";
+import {IHumidityDevice} from "../../../../../server/entities/device.interface";
 
 @Injectable()
 export class HumidityDeviceCacheService extends GenericeCacheService<IHumidityDevice> {
