+++ client/src/app/cache/service/user.cache.service.ts
@@ -1,9 +1,9 @@
 import {Injectable} from "@angular/core";
 import {AuthHttp} from "angular2-jwt";
 import {GenericeCacheService} from "./generic.cache.service";
-import {ClientSocketService} from "../remote/client-socket.service";
-import {NotificationService} from "../notification/notification.service";
-import {IUser} from "../../../../server/entities/user.interface";
+import {ClientSocketService} from "../../remote/client-socket.service";
+import {NotificationService} from "../../notification/notification.service";
+import {IUser} from "../../../../../server/entities/user.interface";
 
 @Injectable()
 export class UserCacheService extends GenericeCacheService<IUser> {
